Extract "Trackers" from IRacingPlugin
Each Tracker tracks data to produce an certain event. E.g
IncidentTracker. It's pretty much just the HandleXXX() methods that have
been moved into a class. They all share a TrackerState, which contains
the values they use to compare with previous reported state
