+++ Components/IRacing/Plugins/Models/DriverState.cs
@@ -1,0 +1,12 @@
+﻿namespace Slipstream.Components.IRacing.Plugins.Models
+{
+    internal class DriverState
+    {
+        public int PlayerCarDriverIncidentCount { get; set; }
+
+        public void ClearState()
+        {
+            PlayerCarDriverIncidentCount = 0;
+        }
+    }
+}
