+++ Slipstream.csproj
@@ -162,6 +162,20 @@
     <Compile Include="Components\Internal\LuaGLues\CoreLuaGlueFactory.cs" />
     <Compile Include="Components\Internal\Internal.cs" />
     <Compile Include="Components\IRacing\LuaGlueFactory.cs" />
+    <Compile Include="Components\IRacing\Plugins\Models\CarState.cs" />
+    <Compile Include="Components\IRacing\Plugins\Models\DriverState.cs" />
+    <Compile Include="Components\IRacing\Plugins\Trackers\CarInfoTracker.cs" />
+    <Compile Include="Components\IRacing\Plugins\Trackers\CarPositionTracker.cs" />
+    <Compile Include="Components\IRacing\Plugins\Trackers\IRacingSessionTracker.cs" />
+    <Compile Include="Components\IRacing\Plugins\Trackers\PitUsageTracker.cs" />
+    <Compile Include="Components\IRacing\Plugins\Trackers\LapsCompletedTracker.cs" />
+    <Compile Include="Components\IRacing\Plugins\Trackers\FlagsTracker.cs" />
+    <Compile Include="Components\IRacing\Plugins\Trackers\IncidentTracker.cs" />
+    <Compile Include="Components\IRacing\Plugins\Trackers\IIRacingDataTracker.cs" />
+    <Compile Include="Components\IRacing\Plugins\Trackers\ConnectTracker.cs" />
+    <Compile Include="Components\IRacing\Plugins\Models\IRacingDataTrackerState.cs" />
+    <Compile Include="Components\IRacing\Plugins\Trackers\Trackers.cs" />
+    <Compile Include="Components\IRacing\Plugins\Trackers\WeatherTracker.cs" />
     <Compile Include="Components\Lua\CapturingEventBus.cs" />
     <Compile Include="Components\Playback\LuaGlueFactory.cs" />
     <Compile Include="Components\Playback\Plugins\PlaybackPlugin.cs" />
